+++ Controllers/DriversController.cs
@@ -10,17 +10,21 @@ namespace go_bus_backend.Controllers;
 public class DriversController : ControllerBase
 {
     private readonly IDriverRepository _driverRepository;
-  
+
 
     public DriversController(IDriverRepository driverRepository)
     {
         _driverRepository = driverRepository;
-       
     }
 
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] AddDriverRequestDto addDriverRequestDto)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
         var driver = new Driver()
         {
             Name = addDriverRequestDto.Name,
@@ -29,7 +33,6 @@ public class DriversController : ControllerBase
             DateOfBirth = addDriverRequestDto.DateOfBirth,
             DriverStatus = bool.Parse(addDriverRequestDto.DriverStatus),
             ContactNumber = addDriverRequestDto.ContactNumber,
-            
         };
         var createdDriver = await _driverRepository.CreateAsync(driver);
         return Ok(createdDriver);
@@ -38,9 +41,17 @@ public class DriversController : ControllerBase
     [HttpPut]
     public async Task<IActionResult?> Update(int id, [FromBody] UpdateDriverRequestDto updateDriverRequestDto)
     {
-        
-        
-        var bus = new Driver()
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
+        var oldDriver = await _driverRepository.GetById(id);
+        if (oldDriver == null)
+            return NotFound();
+
+
+        var driver = new Driver()
         {
             Name = updateDriverRequestDto.Name,
             Surname = updateDriverRequestDto.Surname,
@@ -48,15 +59,23 @@ public class DriversController : ControllerBase
             DateOfBirth = updateDriverRequestDto.DateOfBirth,
             DriverStatus = updateDriverRequestDto.DriverStatus,
             ContactNumber = updateDriverRequestDto.ContactNumber,
-
         };
-        var updatedBus = await _driverRepository.UpdateAsync(id, bus);
-        return Ok(updatedBus);
+        var updatedDriver = await _driverRepository.UpdateAsync(id, driver);
+        return Ok(updatedDriver);
     }
 
     [HttpDelete]
     public async Task<IActionResult> Delete([FromBody] int id)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
+        var driver = await _driverRepository.GetById(id);
+        if (driver == null)
+            return NotFound();
+
         var deletedDriver = await _driverRepository.DeleteAsync(id);
         return Ok(deletedDriver);
     }
@@ -66,6 +85,11 @@ public class DriversController : ControllerBase
         [FromQuery] string? sortBy, [FromQuery] bool? isAscending, [FromQuery] int page = 1,
         [FromQuery] int pageSize = 10)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
         var drivers = await _driverRepository.GetAllAsync(filterOn, filterQuery, sortBy, isAscending ?? true, page,
             pageSize);
         return Ok(drivers);
@@ -74,7 +98,12 @@ public class DriversController : ControllerBase
     [HttpGet("GetById")]
     public async Task<IActionResult> GetById(int id)
     {
-        var bus = await _driverRepository.GetById(id);
-        return Ok(bus);
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
+        var driver = await _driverRepository.GetById(id);
+        return Ok(driver);
     }
 }
